Create code-student-register
This is a program for recording student attendance and hours worked.
It uses a Python SQLite database, where student records are stored and managed.
With tkinter's graphical interface and an Azure-TCL extension you can work in an easy, practical and user-friendly way. 
Students must have a mandatory amount of hours per week, so if they work more than the required hours, they are saved in accumulated hours, this is so they can transfer the hours to other weeks and do not have to attend more days and hours.